+++ common/src/main/kotlin/com/cobblemon/mod/common/api/events/pokemon/FossilRevivedEvent.kt
@@ -1,0 +1,17 @@
+/*
+ * Copyright (C) 2023 Cobblemon Contributors
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ */
+
+package com.cobblemon.mod.common.api.events.pokemon
+
+import com.cobblemon.mod.common.pokemon.Pokemon
+import net.minecraft.server.network.ServerPlayerEntity
+
+data class FossilRevivedEvent (
+    val pokemon: Pokemon,
+    val player: ServerPlayerEntity?
+)